Unify full + incremental pipelines on shared nested-pipeline core
The full-scan and incremental-append pipelines previously had two parallel
implementations that mirrored each other in shape: separate
FuturesUnordered orchestrators, separate spawn_*_file_task functions, and
separate process_file loops. They are now one implementation reached from
two thin entry points.

Key changes:

* Shared core in nested_pipeline.rs (renamed from ordered_file_pipeline.rs
  via git mv):
    - create_nested_pipeline<T, Src, F> — generic orchestrator on top of
      Stream::buffered(prefetch_depth), replacing the FuturesUnordered
      driver. Cap on alive process_file tasks is now exactly
      prefetch_depth + (files held by FFI).
    - spawn_file_task<S, BSF> — single per-file setup + spawn, parameterized
      by a build_batch_stream closure.
    - process_file<S, BSF> — single decode/serialize/send loop.

* Two-tier per-file slot semaphore: UNATTACHED_SLOTS=1 caps prefetch ahead
  of FFI handoff; promoted to ATTACHED_SLOTS=8 in
  IcebergFileScanResponse::set_payload (and inlined in the flat
  create_pipeline wrapper, which bypasses the handoff).

* full_pipeline.rs (new) hosts create_full_scan_pipeline (nested FFI) and
  create_pipeline (legacy flat FFI wrapper = nested + try_flatten with
  inline slot promotion) plus the real-parquet end-to-end tests.

* incremental_pipeline.rs slims down to read_one_append_file + the
  create_incremental_nested_pipeline glue, calling into the shared core.
  Three stats counters (reader_setup_ns, peak_concurrency,
  buffered_bytes) are now correctly populated on this path — they were
  silently zero before because the parallel implementation didn't write
  them.

* Single concurrency knob: file_prefetch_depth. The previous
  data_file_concurrency_limit was a separate cap that overlapped
  confusingly with prefetch_depth; the FFI setter
  iceberg_scan_with_data_file_concurrency_limit and its incremental twin
  are now no-op stubs (kept for ABI compatibility). The dead
  file_concurrency fields on IcebergScan / IcebergIncrementalScan are
  kept for the same reason and for shape-symmetry with the scan_common.rs
  macros.

* pipeline_stats.rs: wall-clock is now timestamp-based
  (pipeline_start_ns captured at STATS.reset(), pipeline_end_ns via
  fetch_max on per-file stream close), replacing the previous
  pipeline_wall_ns duration that disagreed between flat and nested
  paths. Producer-side waits collapse into one counter
  (producer_stall_ns) covering byte_sem + slot_sem + tx.send.

* Tests: orchestration tests in nested_pipeline.rs use a stub start_one
  closure for fast plumbing checks; real-parquet tests in
  full_pipeline.rs and incremental_pipeline.rs exercise the full per-file
  path. A pub(crate) PIPELINE_TEST_LOCK serializes tests that touch the
  global STATS so concurrent cargo-test workers don't race.

* Design doc: iceberg_rust_ffi/docs/pipeline-unification.md describes the
  net state with before/after diagrams.
